feat(combinatorics/simple_graph/connectivity): walks, paths, cycles
This is the first chunk of #8737, which gives a type for walks in a
simple graph as well as some basic operations.

It is designed to one day generalize to other types of graphs once there
is a more generic framework by swapping out the `G.adj u v` argument
from `walk`.