refactor(combinatorics/quiver): split into several files (#11006)
This PR splits `combinatorics/quiver.lean` into several files. The main reason for this is to ensure that the category theory library only imports the necessary definitions (and not for example the stuff on arborescences).

Also adds some more documentation, and fixes a bug in the definition of weakly connected components.