feat(data/complex/module): `conj_ae` matrix representation and determinant
Add lemmas giving the matrix representation of `conj_ae` (with respect
to `basis_one_I`), and its determinant (both as a linear map and as a
linear equiv).  This is preparation for doing things about how
isometries affect orientations (so it's actually `conj_lie` I'm
interested in, but `conj_lie` is defined in terms of `conj_ae` so
proving things first for `conj_ae` seems appropriate).

---

Note: I realise it may not be desirable to add the import of
`linear_algebra.determinant` to a fairly low-level file such as
`data.complex.module`, I'm just not sure where best to put the two
lemmas about determinants if not in this file.  A separate file such
as `data.complex.determinant` just for those two lemmas?