feat(analysis/complex/isometry): `rotation` matrix representation and determinant
Add lemmas giving the matrix representation of `rotation` (with
respect to `basis_one_I`), and its determinant (both as a linear map
and as a linear equiv).  This is preparation for doing things about
how isometries affect orientations.

---

As with #11337, there is the question of whether it's desirable to add
an import of `linear_algebra.determinant` here, and, if not, where to
put the two lemmas about determinants, though maybe adding such an
import here is of less concern than adding it to
`data.complex.module`.