refactor(linear_algebra/{tensor,exterior}_algebra): convert to a directory
These files can be quite slow, so it's useful to be able to add new definitions and lemmas in standalone files, rather than in the same file.
There are a number of open PRs that make this move as part of a new feature, so let's just move them pre-emptively.