feat(analysis/normed_space/linear_isometry): `to_linear_equiv_trans`
Add a lemma relating `trans` for `linear_isometry_equiv` and
`linear_equiv`.

---

Note: I'm not sure which way round this should be a `simp` lemma.
This version is the same way round as `ring_equiv.to_ring_hom_trans`,
for example, but `linear_equiv.comp_coe` is the other way round.