feat(field_theory/is_alg_closed/classification): Classify algebraically closed fields (#9370)
The main results here are that two algebraically closed fields with the same characteristic and the same cardinality of transcendence basis are isomorphic. The consequence of this is that two uncountable algebraically closed fields of the same cardinality and characteristic are isomorphic. This has applications in model theory, in particular the Lefschetz principle https://proofwiki.org/wiki/Lefschetz_Principle_(First-Order)



Co-authored-by: Chris Hughes <33847686+ChrisHughes24@users.noreply.github.com>