feat(geometry/euclidean/oriented_angle): oriented angles and rotations
Define oriented angles and rotations in a real inner product space,
with respect to a choice of orthonormal basis indexed by `fin 2`, and
prove various lemmas about them, including that the definition depends
only on the orientation associated with the basis, and geometrical
results such as pons asinorum, "angle at center of a circle equals
twice angle at circumference" and "angles in same segment are equal" /
"opposite angles of a cyclic quadrilateral add to π" (these last two
being the same result for oriented angles mod π).

---

There are certainly various additional lemmas that could be added, but
this seems a reasonable starting point for review.

In followup PRs I intend to add corresponding API that takes an
`orientation` instead of an orthonormal basis (defined using
`orientation.fin_orthonormal_basis` and with each lemma deduced from
the corresponding one taking an orthonormal basis); that API, not the
one involving a choice of basis, is expected to be the main API for
users of oriented angles between vectors (hence the lack of a "Main
definitions" section in the module doc string; none of the definitions
added in this PR are intended to be the main definitions for users);
but it seems a good idea to complete review of the API in this PR
before adding a copy of most of it that takes an `orientation`.  I
also intend to add corresponding API (and notation) in Euclidean
affine spaces, taking the orientation from a `module.oriented`
instance.  (Both of those would probably also use a `fact` instance
for the space being two-dimensional.)

This PR also has nothing about the relation between oriented and
unoriented angles; I intend to include that in followups as well.

Once there is a sufficiently fully-featured API for a bundled
`orthonormal_basis` (including coercions to `basis`, analogues of
`basis.map` and `basis.equiv` with lemmas that they are consistent
with the corresponding `basis` operations, and
`analysis.inner_product_space.orientation` and various other lemmas
updated to use that API), it will be appropriate to switch everything
here that uses a `basis` that is `orthonormal` over to the bundled
type.