feat(analysis/inner_product_space/pi_L2): `linear_isometry.extend`
Let `S` be a subspace of a finite-dimensional complex inner product
space `V`.  A linear isometry mapping `S` into `V` can be extended to a
full isometry of `V`.  Note that this is false if we remove either the
finite-dimensional or the complex hypotheses.

I hope that the naming choice is consistent.  Combining the two
`simp only` blocks in `linear_isometry.extend_apply` results in a
timeout, but they seem to work okay split up.