feat(geometry/euclidean/oriented_angle): oriented angles with respect to an orientation
Add definitions and lemmas for oriented angles defined to take an
orientation, instead of an orthonormal basis, as an argument.  These
are the versions intended to be used by most users when working with
oriented angles between vectors, instead of users needing to deal with
a choice of basis.

Apart from the last five lemmas that relate angles and rotations for
different orientations or relate them explicitly to the definitions
with respect to a basis, everything is deduced directly from the
corresponding lemma that takes an orthonormal basis as an argument.

---

A further followup is intended to add a corresponding API (and
notation) in a Euclidean affine space, but that won't have such an
exact one-to-one correspondence of lemmas.