feat(analysis/inner_product_space/sqrt): `is_self_adjoint.sqrt`
We define the square root of a self-adjoint operator `T` on a
finite-dimensional inner product space `E` in terms of an orthonormal
basis of eigenvectors for `T`.

Once `analysis.inner_product_space.spectrum` contains spectral theory
for bounded operators, we can generalize to bounded self-adjoint
operators.  In the meantime, this finite-dimensional result will be
useful for polar and singular value decompositions.