feat(topology/bornology/basic): Define bornology (#12036)
This defines bornologies and provides the basic API. A bornology on a type is a filter consisting of cobounded sets which contains the cofinite sets; bounded sets are then complements of cobounded sets. This is equivalent to the usual formulation in terms of bounded sets, but provides access to mathlib's large filter library. We provide the relevant API for bounded sets.



Co-authored-by: YaelDillies <yael.dillies@gmail.com>