feat(algebra/square): Add the notion of "square" elements of monoids
There is already a concept of `even` and `odd`, but not of `square`.
These are some initial lemmas about being a square that I use in the
proof of the density of squarefree integers.

Note that this also introduces the notation `sqrt0`, which is `0` if its
argument is not a square and otherwise is its squareroot. This is in
line with `real.sqrt`'s behavior on negative numbers. However, the name
perhaps could use some work?